Fixed bugs

+++ Coursera/max_pairwise_product.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
-
+#include <stdint.h>
 int main()
 {
-	int n , n1;
-	scanf("%d",&n);
-	int list_of_numbers[n], largest_num, second_largest_num;
+	int64_t n , n1;
+	scanf("%ld",&n);
+	int64_t list_of_numbers[n], largest_num, second_largest_num;
 	for(int i =0; i<n; i++)
 	{
-		scanf("%d",(list_of_numbers + i));
+		scanf("%ld",(list_of_numbers + i));
 	}
 	largest_num = *(list_of_numbers);
 	for(int i =1; i<n; i++)
@@ -25,7 +25,7 @@ int main()
 			second_largest_num = *(list_of_numbers+i);
 		}
 	}
-	printf("%d", (largest_num * second_largest_num));
+	printf("%ld", (largest_num * second_largest_num));
 
 	return (0);
 }
